PROGRESS/REFACTOR: Added data validation; added ApiResponse class, which is wired as the response object for all requests

+++ FinSharkApi/FinShark.API/Controllers/CommentController.cs
@@ -1,6 +1,9 @@
-﻿using FinShark.API.DTOs.Comment;
+﻿using FinShark.API.DTOs;
+using FinShark.API.DTOs.Comment;
+using FinShark.API.Helpers;
 using FinShark.API.Interfaces;
 using FinShark.API.Mappers;
+using FinShark.API.Models;
 using Microsoft.AspNetCore.Mvc;
 
 namespace FinShark.API.Controllers;
@@ -21,25 +24,46 @@ public class CommentController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if (!ModelState.IsValid)
+        {
+            var errors = ValidationHelpers.GetValidationErrors(ModelState);
+            return BadRequest(ApiResponse<Dictionary<string, List<string>>>
+                .FailureResponse("Invalid validation", errors));
+        }
+
         var comments = await _commentRepo.GetAllAsync();
         var commentDtos = comments.Select(c => c.ToCommentDto()).ToList();
-        return Ok(commentDtos);
+        return Ok(ApiResponse<List<CommentDto>>.SuccessResponse(commentDtos));
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+        {
+            var errors = ValidationHelpers.GetValidationErrors(ModelState);
+            return BadRequest(ApiResponse<Dictionary<string, List<string>>>
+                .FailureResponse("Invalid validation", errors));
+        }
+
         var comment = await _commentRepo.GetByIdAsync(id);
         return comment != null ?
-            Ok(comment.ToCommentDto()) :
-            NotFound();
+            Ok(ApiResponse<CommentDto>.SuccessResponse(comment.ToCommentDto())) :
+            NotFound(ApiResponse<CommentDto>.FailureResponse("The comment does not exist."));
     }
 
-    [HttpPost("{stockId}")]
+    [HttpPost("{stockId:int}")]
     public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto commentDto)
     {
+        if (!ModelState.IsValid)
+        {
+            var errors = ValidationHelpers.GetValidationErrors(ModelState);
+            return BadRequest(ApiResponse<Dictionary<string, List<string>>>
+                .FailureResponse("Invalid validation", errors));
+        }
+
         if (!await _stockRepo.StockExists(stockId))
-            return BadRequest("The stock doesn't exist.");
+            return BadRequest(ApiResponse<CommentDto>.FailureResponse("The stock does not exist."));
 
         var commentModel = commentDto.ToCommentFromCreateDto(stockId);
         await _commentRepo.CreateAsync(commentModel);
@@ -47,22 +71,36 @@ public class CommentController : ControllerBase
     }
 
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateDto)
     {
+        if (!ModelState.IsValid)
+        {
+            var errors = ValidationHelpers.GetValidationErrors(ModelState);
+            return BadRequest(ApiResponse<Dictionary<string, List<string>>>
+                .FailureResponse("Invalid validation", errors));
+        }
+
         var commentModel = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdateDto());
         return commentModel != null ?
-            Ok(commentModel.ToCommentDto()) :
-            NotFound();
+            Ok(ApiResponse<CommentDto>.SuccessResponse(commentModel.ToCommentDto())) :
+            NotFound(ApiResponse<CommentDto>.FailureResponse("The comment does not exist."));
     }
 
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+        {
+            var errors = ValidationHelpers.GetValidationErrors(ModelState);
+            return BadRequest(ApiResponse<Dictionary<string, List<string>>>
+                .FailureResponse("Invalid validation", errors));
+        }
+
         var commentModel = await _commentRepo.DeleteAsync(id);
         return commentModel != null ?
             NoContent() :
-            NotFound();
+            NotFound(ApiResponse<Comment>.FailureResponse("The comment does not exist."));
     }
 }
